refactor: handlers and routes
