+++ CHANGELOG.md
@@ -1,0 +1,9 @@
+# Changes
+
+## v0.0.9
+
+### Added
+
+* [CLI](https://github.com/EstebanBorai/http-server/commit/f7ea5b278e8b46c322d82c324c7f263ecae3914b)
+* [Config Shape](https://github.com/EstebanBorai/http-server/commit/49e4cd73ddc4401bf684ad70875b075bb35caf5a)
+* [HTTP server with File Explorer](https://github.com/EstebanBorai/http-server/commit/7aa59495dab51c52a0a79b75c185b3b74dda5be1)
