+++ src/handler/mod.rs
@@ -1,3 +1,5 @@
-mod static_fs;
+mod file_explorer;
+mod main_handler;
 
-pub use static_fs::*;
+pub use file_explorer::*;
+pub use main_handler::*;
